[NQN] Fixed to send available emojis properly

+++ src/events/messageCreate.ts
@@ -1,25 +1,40 @@
 import {Client, Message, TextChannel, Webhook} from 'discord.js';
-//import {getMap} from '../util/emojiMap';
+import {getMap} from '../util/emojiMap';
+const regex = new RegExp(String.raw`<a?:\w+:\d+>|(?<!\\):(\w+):`, 'g')
+import * as _ from 'lodash';
 
 module.exports = async (client: Client, message: Message) => {
-    
-    if (message.content.toLowerCase().includes(":baldmeltypet:")) {
 
-        if (message.guild?.me?.permissions.has("MANAGE_MESSAGES")) {
-            message.delete()
-        }
-        
-        //const emojiMap = getMap();
+    // REPLACE EMOJI
+    if (message.author.bot) return;
+    let emojis = message.content.match(regex);
 
-        //const newMsg = message.content.replace(/:baldmeltypet:/gi, emojiMap.get('baldmeltypet'))
-        const newMsg = message.content;
-        const channel = message.channel as TextChannel;
-        const webhooks = await channel.fetchWebhooks();
-        let webhook = webhooks.find((w) => w.token != null);
+    if (emojis == null) return;
 
-        if (!webhook) {
-            webhook = await channel.createWebhook('Bald Macaron', {
-            avatar: client.user?.displayAvatarURL(),
+    emojis = _.uniq(emojis);
+    const emojiMap = getMap();
+
+    let newMsg = message.content;
+
+    emojis!.forEach(async em => {
+        const emoji = emojiMap.get(em)
+        if (!emoji) return;
+        newMsg = newMsg.replaceAll(em, emojiMap.get(em))
+    });
+
+
+    // send emojis
+    if (newMsg == message.content) return;
+    if (message.guild?.me?.permissions.has("MANAGE_MESSAGES")) {message.delete()}
+
+
+    const channel = message.channel as TextChannel;
+    const webhooks = await channel.fetchWebhooks();
+    let webhook = webhooks.find((w) => w.token != null);
+
+    if (!webhook) {
+        webhook = await channel.createWebhook('Bald Macaron', {
+            avatar: client.user?.avatarURL(),
         });
     }
 
@@ -28,6 +43,5 @@ module.exports = async (client: Client, message: Message) => {
         username: message.author.username,
         content: newMsg,
     });
-
-    }
 }
+
